# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Intention app",
    "description": "Business logic of Intention App",
    "termsOfService": "N/A",
    "license": {
      "name": "Proprietary, not free to use",
      "url": "N/A"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": ""
    }
  ],
...
  "paths": {
...
    "/api/v1/projects/{intentionId}": {
      "get": {
        "tags": ["project-controller"],
        "summary": "Get all projects by intention id",
        "operationId": "getProjects",
        "parameters": [
          {
            "name": "intentionId",
            "in": "path",
            "description": "Id of intention",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "description": "Number of page(min: 1)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "itemsPerPage",
            "in": "query",
            "description": "Items per page(min: 1, max: 100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sorting",
            "in": "query",
            "description": "Sorting by createdAt",
            "allowEmptyValue": true,
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User projects",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectContentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request, invalid request data",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "example": {
                    "success": false,
                    "errors": [
                      {
                        "value": "0",
                        "msg": "Invalid value",
                        "param": "pageNumber",
                        "location": "query"
                      }
                    ]
                  }
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "example": {
                    "success": false
                  }
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "Bearer <JWT>": []
          }
        ],
        "x-codegen-request-body-name": "description"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProjectContentResponse": {
        "title": "ProjectContentResponse",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "subtitle": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "subprojects": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer <JWT>": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}